fix: unit tests work with typescript
