+++ lib/contentful-management.js
@@ -5,7 +5,7 @@
  * @see ContentfulClientAPI
  */
 import createContentfulApi from './create-contentful-api'
-import { createHttpClient } from './create-http-client'
+import { createCMAHttpClient } from './create-cma-http-client'
 
 /**
  * Create a client instance
@@ -20,7 +20,7 @@ import { createHttpClient } from './create-http-client'
  * })
  */
 export function createClient (params) {
-  const http = createHttpClient(params)
+  const http = createCMAHttpClient(params)
   const api = createContentfulApi({
     http: http
   })
