+++ package.json
@@ -33,10 +33,10 @@
     "pretest": "rimraf coverage && npm run lint",
     "test": "npm run test:cover && npm run test:integration && npm run test:browser-local && npm run test:size",
     "test:ci": "./node_modules/contentful-sdk-core/bin/test-ci.sh",
-    "test:cover": "BABEL_ENV=test babel-node ./node_modules/istanbul/lib/cli.js cover ./test/runner | tap-spec",
-    "test:unit": "BABEL_ENV=test babel-node ./test/runner | tap-spec",
-    "test:debug": "BABEL_ENV=test babel-node debug ./test/runner",
-    "test:integration": "BABEL_ENV=test babel-node ./test/integration/integration-tests.js",
+    "test:cover": "BABEL_ENV=test babel-node  --extensions \".ts\" --extensions \".js\" ./node_modules/istanbul/lib/cli.js cover ./test/runner.js | tap-spec",
+    "test:unit": "BABEL_ENV=test babel-node --extensions \".ts\" --extensions \".js\" ./test/runner.js | tap-spec",
+    "test:debug": "BABEL_ENV=test babel-node debug  --extensions \".ts\" --extensions \".js\"  ./test/runner.js",
+    "test:integration": "BABEL_ENV=test babel-node  --extensions \".ts\" --extensions \".js\"  ./test/integration/integration-tests.js",
     "test:browser-local": "BABEL_ENV=test karma start karma.conf.local.js",
     "test:browser-remote": "BABEL_ENV=test karma start karma.conf.saucelabs.js",
     "test:e2e": "node test/e2e/index.js",
@@ -69,14 +69,14 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-typescript": "^7.9.0",
-    "@types/lodash": "^4.14.152",
     "@semantic-release/changelog": "^5.0.1",
+    "@types/lodash": "^4.14.152",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "babel-minify-webpack-plugin": "^0.3.1",
     "babel-plugin-inline-replace-variables": "^1.3.1",
     "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-rewire": "^1.2.0",
+    "babel-plugin-rewire-ts": "1.3.3",
     "blue-tape": "^1.0.0",
     "bundlesize": "^0.18.0",
     "codecov": "^3.1.0",
