+++ typings/contentfulManagement.d.ts
@@ -1,5 +1,5 @@
 import { AxiosRequestConfig, AxiosResponse } from 'axios'
-import { ClientParams } from './create-http-client';
+import { ClientParams } from './create-cma-http-client';
 import { User } from './user'
 import { PersonalAccessToken, PersonalAccessTokenProp } from './personalAccessToken'
 import { Space, SpaceProps } from './space'
@@ -7,7 +7,7 @@ import { Collection } from './collection'
 import { Organization } from './organization'
 import * as Usage from './usage'
 
-export { ClientParams } from './create-http-client'
+export { ClientParams } from './create-cma-http-client'
 
 export as namespace contentfulManagementStatic
